requirements.txt
Django==3.1.6
black==22.1.0
gunicorn==20.1.0
psycopg2-binary==2.9.1
python-dotenv==0.19.2
sentry-sdk==1.5.10
raven==6.10.0
Pillow==9.1.0
pyjwt==2.3.0
ipython==8.2.0
